# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Google Ads Search Query Analysis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RuleSet {
  id: number;
  name: string;
  /** @nullable */
  accountName?: string | null;
  competitorBrands?: string[];
  excludePatterns?: string[];
  customRules?: string[];
  minConversionsForNewKeyword?: number;
  /**
     * Saved active keywords for this account profile
     * @nullable
     */
  activeKeywords?: string | null;
  /**
     * Saved landing page URL for this account profile
     * @nullable
     */
  landingPageUrl?: string | null;
  /**
     * Saved target service locations for this account profile
     * @nullable
     */
  targetLocations?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface RuleSetInput {
  name: string;
  accountName?: string;
  competitorBrands?: string[];
  excludePatterns?: string[];
  customRules?: string[];
  minConversionsForNewKeyword?: number;
  /** Active keywords to save with this account profile */
  activeKeywords?: string;
  /** Landing page URL to save with this account profile */
  landingPageUrl?: string;
  /** Target service locations to save with this account profile */
  targetLocations?: string;
}

export interface AnalysisInput {
  name?: string;
  /** @nullable */
  ruleSetId?: number | null;
  /** Optional URL of the landing page/website to scan for relevance context */
  landingPageUrl?: string;
  /** Optional comma-separated target service locations (cities, regions). Terms outside these areas will be flagged as irrelevant. */
  targetLocations?: string;
  /** Newline or comma separated list of active keywords with their match types and ad groups */
  activeKeywords: string;
  /** Newline or tab-separated search term data (term, impressions, clicks, conversions, cost) */
  searchTerms: string;
  competitorBrands?: string[];
  excludePatterns?: string[];
  customRules?: string[];
  minConversionsForNewKeyword?: number;
}

export type SearchTermResultRelevance = typeof SearchTermResultRelevance[keyof typeof SearchTermResultRelevance];


export const SearchTermResultRelevance = {
  Relevant: 'Relevant',
  Irrelevant: 'Irrelevant',
} as const;

export type SearchTermResultAddLevel = typeof SearchTermResultAddLevel[keyof typeof SearchTermResultAddLevel];


export const SearchTermResultAddLevel = {
  Campaign: 'Campaign',
  Ad_Group: 'Ad Group',
  None: 'None',
} as const;

export interface SearchTermResult {
  searchTerm: string;
  /** @nullable */
  impressions?: number | null;
  /** @nullable */
  clicks?: number | null;
  /** @nullable */
  conversions?: number | null;
  /** @nullable */
  cost?: number | null;
  relevance: SearchTermResultRelevance;
  reason: string;
  addLevel: SearchTermResultAddLevel;
  addAsKeyword: boolean;
  /** @nullable */
  suggestedAdGroup: string | null;
  isCompetitor: boolean;
  /** @nullable */
  matchedKeyword?: string | null;
  /**
     * The specific location word/phrase in the search term that is outside the target service area
     * @nullable
     */
  outOfAreaLocation?: string | null;
}

export type AnalysisStatus = typeof AnalysisStatus[keyof typeof AnalysisStatus];


export const AnalysisStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Analysis {
  id: number;
  name: string;
  status: AnalysisStatus;
  /** @nullable */
  ruleSetId?: number | null;
  /** The raw active keywords string used for this analysis */
  activeKeywords?: string;
  /**
     * Target service locations used for this analysis
     * @nullable
     */
  targetLocations?: string | null;
  /**
     * Landing page URL used for this analysis
     * @nullable
     */
  landingPageUrl?: string | null;
  competitorBrands?: string[];
  excludePatterns?: string[];
  customRules?: string[];
  minConversionsForNewKeyword?: number;
  totalTerms?: number;
  relevantCount?: number;
  irrelevantCount?: number;
  newKeywordCount?: number;
  competitorCount?: number;
  results: SearchTermResult[];
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}

export type AnalysisSummaryStatus = typeof AnalysisSummaryStatus[keyof typeof AnalysisSummaryStatus];


export const AnalysisSummaryStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface AnalysisSummary {
  id: number;
  name: string;
  status: AnalysisSummaryStatus;
  /** @nullable */
  ruleSetId?: number | null;
  totalTerms: number;
  relevantCount: number;
  irrelevantCount: number;
  newKeywordCount?: number;
  competitorCount?: number;
  createdAt: string;
}

export interface AnalysisStats {
  totalAnalyses: number;
  totalTermsAnalyzed: number;
  avgRelevanceRate: number;
  topCompetitors: string[];
}