# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Google Ads Search Query Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisStatus } from './analysisStatus';
import type { SearchTermResult } from './searchTermResult';

export interface Analysis {
  id: number;
  name: string;
  status: AnalysisStatus;
  /** @nullable */
  ruleSetId?: number | null;
  /** The raw active keywords string used for this analysis */
  activeKeywords?: string;
  /**
     * Target service locations used for this analysis
     * @nullable
     */
  targetLocations?: string | null;
  /**
     * Landing page URL used for this analysis
     * @nullable
     */
  landingPageUrl?: string | null;
  competitorBrands?: string[];
  excludePatterns?: string[];
  customRules?: string[];
  minConversionsForNewKeyword?: number;
  totalTerms?: number;
  relevantCount?: number;
  irrelevantCount?: number;
  newKeywordCount?: number;
  competitorCount?: number;
  results: SearchTermResult[];
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}