# lib/api-zod/src/generated/types/ruleSet.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Google Ads Search Query Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface RuleSet {
  id: number;
  name: string;
  /** @nullable */
  accountName?: string | null;
  competitorBrands?: string[];
  excludePatterns?: string[];
  customRules?: string[];
  minConversionsForNewKeyword?: number;
  /**
     * Saved active keywords for this account profile
     * @nullable
     */
  activeKeywords?: string | null;
  /**
     * Saved landing page URL for this account profile
     * @nullable
     */
  landingPageUrl?: string | null;
  /**
     * Saved target service locations for this account profile
     * @nullable
     */
  targetLocations?: string | null;
  createdAt: string;
  updatedAt: string;
}