# lib/api-zod/src/generated/types/ruleSetInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Google Ads Search Query Analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface RuleSetInput {
  name: string;
  accountName?: string;
  competitorBrands?: string[];
  excludePatterns?: string[];
  customRules?: string[];
  minConversionsForNewKeyword?: number;
  /** Active keywords to save with this account profile */
  activeKeywords?: string;
  /** Landing page URL to save with this account profile */
  landingPageUrl?: string;
  /** Target service locations to save with this account profile */
  targetLocations?: string;
}