# lib/api-zod/src/generated/types/searchTermResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Google Ads Search Query Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchTermResultAddLevel } from './searchTermResultAddLevel';
import type { SearchTermResultRelevance } from './searchTermResultRelevance';

export interface SearchTermResult {
  searchTerm: string;
  /** @nullable */
  impressions?: number | null;
  /** @nullable */
  clicks?: number | null;
  /** @nullable */
  conversions?: number | null;
  /** @nullable */
  cost?: number | null;
  relevance: SearchTermResultRelevance;
  reason: string;
  addLevel: SearchTermResultAddLevel;
  addAsKeyword: boolean;
  /** @nullable */
  suggestedAdGroup: string | null;
  isCompetitor: boolean;
  /** @nullable */
  matchedKeyword?: string | null;
  /**
     * The specific location word/phrase in the search term that is outside the target service area
     * @nullable
     */
  outOfAreaLocation?: string | null;
}